feat: Introduce the LinearGroupByKey/Mut types
